integrate ngx-toastr & add confirmation messages

+++ src/app/_pages/notes-list/notes-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { NotesService } from '../../_shared/services/notes.service';
 import { Note } from '../../_shared/models/note.model';
 import {Router} from '@angular/router';
+import {ToastrService} from 'ngx-toastr';
 
 @Component({
   selector: 'app-notes-list',
@@ -9,27 +10,47 @@ import {Router} from '@angular/router';
   styleUrls: ['./notes-list.component.scss']
 })
 export class NotesListComponent implements OnInit {
+  loading = false;
   inputSelected = false;
   notes: Note[] = [];
 
   constructor(
     private router: Router,
+    private toast: ToastrService,
     private notesService: NotesService,
   ) { }
 
   ngOnInit(): void {
-    console.log(new Date());
+    this.loading = true;
     if (window.history.state.action !== undefined && window.history.state.result !== undefined) {
       console.log(window.history.state.action, window.history.state.result);
+      switch (window.history.state.result) {
+        case 'success':
+          if (window.history.state.action === 'add') { // add new note
+            this.onSuccess('Note added successfully!', 'Add Note');
+          } else { // update note
+            this.onSuccess('Note updated successfully!', 'Update Note');
+          }
+          break;
+        case 'error':
+          this.onError();
+          break;
+        default:
+          // do nothing
+          break;
+      }
     }
 
     this.notesService.getNotes().subscribe(data => {
+      this.loading = false;
       this.notes = data.map(e => {
         return {
           id: e.payload.doc.id,
           ...e.payload.doc.data()
         } as Note;
       });
+    }, err => {
+      this.onError();
     });
   }
 
@@ -42,7 +63,23 @@ export class NotesListComponent implements OnInit {
   }
 
   deleteNote(noteId: any): void {
-    this.notesService.deleteNote(noteId);
+    if (confirm('Are you sure ?')) {
+      this.notesService.deleteNote(noteId)
+        .then(r => {
+          this.onSuccess('Note deleted successfully!', 'Delete Note');
+        })
+        .catch(err => {
+          this.onError();
+        });
+    }
+  }
+
+  onSuccess(message: string, title: string): void {
+    this.toast.success(message, title);
+  }
+
+  onError(): void {
+    this.toast.error('Sorry, try later!', 'Error');
   }
 }
